introduce highscore-wrapper
Because Highscore is DBusActivatable,
share/dbus-1/services/app.drey.Highscore.service is executed instead of
share/applications/app.drey.Highscore.desktop under GNOME. However, Exec
in dbus service needs to be an absolute path. We have no choices but
make a wrapper to modify it.

Since we're using a wrapper anyway, we can set environment variables in
the wrapper. This approach is simpler and more upstream-friendly then
the previous argv[0] one.